UCL-454 Moved useMultiSelect to correct location; Enhanced useMultiSelect to automatically generate option elements
